Added Mount Idempotency for block device

+++ pkg/driver/node.go
@@ -586,15 +586,55 @@ func (d *nodeService) nodePublishVolumeForBlock(req *csi.NodePublishVolumeReques
 		return status.Errorf(codes.Internal, "Could not create file %q: %v", target, err)
 	}
 
-	klog.V(4).Infof("NodePublishVolume [block]: mounting %s at %s", source, target)
-	if err := d.mounter.Mount(source, target, "", mountOptions); err != nil {
-		if removeErr := os.Remove(target); removeErr != nil {
-			return status.Errorf(codes.Internal, "Could not remove mount target %q: %v", target, removeErr)
-		}
+	//Checking if the target file is already mounted with a device.
+	mounted, err := d.isMounted(source, target)
+	if err != nil {
 		return status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
 	}
 
+	if !mounted {
+		klog.V(4).Infof("NodePublishVolume [block]: mounting %s at %s", source, target)
+		if err := d.mounter.Mount(source, target, "", mountOptions); err != nil {
+			if removeErr := os.Remove(target); removeErr != nil {
+				return status.Errorf(codes.Internal, "Could not remove mount target %q: %v", target, removeErr)
+			}
+			return status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
+		}
+	} else {
+		klog.V(4).Infof("NodePublishVolume [block]: Target path %q is already mounted", target)
+	}
+
 	return nil
+}
+
+func (d *nodeService) isMounted(source string, target string) (bool, error) {
+	/*
+		Checking if it's a mount point using IsLikelyNotMountPoint. There are three different return values,
+		1. true, err when the directory does not exist or corrupted.
+		2. false, nil when the path is already mounted with a device.
+		3. true, nil when the path is not mounted with any device.
+	*/
+	notMnt, err := d.mounter.IsLikelyNotMountPoint(target)
+	if err != nil && !os.IsNotExist(err) {
+		//Checking if the path exists and error is related to Corrupted Mount, in that case, the system could unmount and mount.
+		_, pathErr := d.mounter.PathExists(target)
+		if pathErr != nil && d.mounter.IsCorruptedMnt(pathErr) {
+			klog.V(4).Infof("NodePublishVolume: Target path %q is a corrupted mount. Trying to unmount.", target)
+			if mntErr := d.mounter.Unmount(target); mntErr != nil {
+				return !notMnt, status.Errorf(codes.Internal, "Unable to unmount the target %q : %v", target, mntErr)
+			}
+			//After successful unmount, the device is ready to be mounted.
+			return !notMnt, nil
+		}
+		return !notMnt, status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
+	}
+
+	if !notMnt {
+		klog.V(4).Infof("NodePublishVolume: Target path %q is already mounted", target)
+		return !notMnt, nil
+	}
+
+	return !notMnt, err
 }
 
 func (d *nodeService) nodePublishVolumeForFileSystem(req *csi.NodePublishVolumeRequest, mountOptions []string, mode *csi.VolumeCapability_Mount) error {
@@ -620,41 +660,20 @@ func (d *nodeService) nodePublishVolumeForFileSystem(req *csi.NodePublishVolumeR
 	mountOptions = collectMountOptions(fsType, mountOptions)
 
 	klog.V(4).Infof("NodePublishVolume: mounting %s at %s with option %s as fstype %s", source, target, mountOptions, fsType)
-	/*
-		Checking if it's a mount point. There are three cases,
-		1. true, err when the directory does not exist or corrupted.
-		2. false, nil when the path is already mounted with a device.
-		3. true, nil when the path is not mounted with any device.
-	*/
-	notMnt, err := d.mounter.IsLikelyNotMountPoint(target)
-	if err != nil && !os.IsNotExist(err) {
-		//Checking if the path exists and error is related to Corrupted Mount, in that case, the system could unmount and mount.
-		_, pathErr := mountutils.PathExists(target)
-		if pathErr != nil && mountutils.IsCorruptedMnt(pathErr) {
-			klog.V(4).Infof("Target path %q is a corrupted directory", target)
-		} else {
+
+	//Checking if the target directory is already mounted with a device.
+	mounted, err := d.isMounted(source, target)
+
+	if err != nil {
+		return status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
+	}
+
+	if !mounted {
+		if err := d.mounter.Mount(source, target, fsType, mountOptions); err != nil {
 			return status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
 		}
 	}
 
-	if !notMnt {
-		//Return error when any of the directory inside is not readable which means that a device could be mounted.
-		_, err = os.ReadDir(target)
-		if err != nil {
-			klog.V(4).Infof("Error occurred at reading directory %q. Trying to Unmount.", target)
-			//Reading the directory failed and trying to unmount and remount
-			if mntErr := d.mounter.Unmount(target); mntErr != nil {
-				return status.Errorf(codes.Internal, "Unable to unmount the target %q : %v", target, err)
-			}
-		} else {
-			klog.V(4).Infof("Target path %q is already mounted", target)
-			return nil
-		}
-	}
-
-	if err := d.mounter.Mount(source, target, fsType, mountOptions); err != nil {
-		return status.Errorf(codes.Internal, "Could not mount %q at %q: %v", source, target, err)
-	}
 	return nil
 }
